# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "LinkSpan REST API",
    "version": "1.0.0",
    "description": "Programmatic access to LinkSpan share links and signaling sessions. Share links upload (client-encrypted) bytes to the server so a recipient can download later, with expiry, password protection, download limits, single-use, and revocation.",
    "license": {
      "name": "MIT"
    }
  },
  "servers": [
    {
      "url": "https://share.example/api/v1"
    }
  ],
  "security": [
    {
      "ApiKeyAuth": []
    },
    {}
  ],
...
  "components": {
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "API key as Bearer token (or X-API-Key header)."
      }
    },
    "schemas": {
...
      "CreateLinkRequest": {
        "type": "object",
        "required": [
          "filename",
          "size"
        ],
        "properties": {
          "filename": {
            "type": "string",
            "maxLength": 255
          },
          "size": {
            "type": "integer",
            "minimum": 0,
            "maximum": 5368709120
          },
          "contentType": {
            "type": "string",
            "default": "application/octet-stream"
          },
          "visibility": {
            "type": "string",
            "enum": [
              "temp",
              "public"
            ],
            "default": "temp"
          },
          "expiresIn": {
            "description": "Preset (5m, 1h, 24h, 7d) or custom milliseconds.",
            "oneOf": [
              {
                "type": "string"
              },
              {
                "type": "integer"
              }
            ]
          },
          "password": {
            "type": "string",
            "maxLength": 256,
            "description": "Optional download password."
          },
          "maxDownloads": {
            "type": "integer",
            "minimum": 1,
            "description": "Multi-use cap; omit for unlimited."
          },
          "singleUse": {
            "type": "boolean",
            "description": "Reaped after the first successful download."
          },
          "metadata": {
            "type": "object",
            "description": "Opaque client metadata (e.g. {\"encrypted\":true})."
          }
        }
      },
...
      "CreateLinkResponse": {
        "allOf": [
          {
            "$ref": "#/components/schemas/ShareLink"
          },
          {
            "type": "object",
            "properties": {
              "uploadToken": {
                "type": "string",
                "description": "Present once; use with PUT .../content."
              },
              "ownerToken": {
                "type": "string",
                "description": "Anonymous capability secret for revoke. Store it."
              },
              "upload": {
                "type": "object"
              }
            }
          }
        ]
      }
    }
  },
  "paths": {
...
    "/telemetry": {
      "post": {
        "tags": [
          "meta"
        ],
        "summary": "Opt-in aggregate telemetry (anonymized, pre-bucketed; no auth)",
        "description": "Clients that have opted in may POST a single anonymized, pre-bucketed transfer event. No filename, byte count, duration value, identity, or per-transfer id is accepted. Always returns 204 (invalid events are silently dropped).",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "outcome": {
                    "type": "string",
                    "enum": [
                      "success",
                      "failure"
                    ]
                  },
                  "mode": {
                    "type": "string",
                    "enum": [
                      "p2p",
                      "relay"
                    ]
                  },
                  "sizeBucket": {
                    "type": "string",
                    "enum": [
                      "lt1mb",
                      "1to10mb",
                      "10to100mb",
                      "100mbto1gb",
                      "gt1gb"
                    ]
                  },
                  "durationBucket": {
                    "type": "string",
                    "enum": [
                      "lt1s",
                      "1to10s",
                      "10to60s",
                      "1to5m",
                      "gt5m"
                    ]
                  }
                },
                "required": [
                  "outcome",
                  "mode",
                  "sizeBucket",
                  "durationBucket"
                ]
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Recorded (or silently dropped if invalid)"
          }
        }
      }
    },
    "/links": {
      "post": {
        "tags": [
          "links"
        ],
        "summary": "Create a share link",
        "security": [
          {
            "ApiKeyAuth": []
          },
          {}
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateLinkRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateLinkResponse"
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Rate limited",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "links"
        ],
        "summary": "List the caller's links",
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "API key required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/links/{id}/content": {
      "put": {
        "tags": [
          "links"
        ],
        "summary": "Upload bytes for a reserved link",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[a-f0-9]{32}$"
            }
          },
          {
            "name": "X-Upload-Token",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/octet-stream": {
              "schema": {
                "type": "string",
                "format": "binary"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Stored",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ShareLink"
                }
              }
            }
          },
          "401": {
            "description": "Bad token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "410": {
            "description": "Expired",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "413": {
            "description": "Too large",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/links/{id}": {
      "get": {
        "tags": [
          "links"
        ],
        "summary": "Link metadata",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[a-f0-9]{32}$"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ShareLink"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "410": {
            "description": "Expired",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "links"
        ],
        "summary": "Revoke a link",
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[a-f0-9]{32}$"
            }
          },
          {
            "name": "X-Owner-Token",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Revoked"
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Not owner",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/links/{id}/download": {
      "get": {
        "tags": [
          "links"
        ],
        "summary": "Download link bytes",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[a-f0-9]{32}$"
            }
          },
          {
            "name": "X-Share-Password",
            "in": "header",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "password",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Bytes",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          },
          "401": {
            "description": "Password required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Wrong password",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "410": {
            "description": "Expired/limit reached",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Rate limited",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/sessions": {
      "post": {
        "tags": [
          "sessions"
        ],
        "summary": "Create a signaling session",
        "security": [
          {
            "ApiKeyAuth": []
          },
          {}
        ],
        "responses": {
          "201": {
            "description": "Created"
          }
        }
      }
    },
    "/sessions/{id}": {
      "get": {
        "tags": [
          "sessions"
        ],
        "summary": "Session status",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    }
  }
}